Issue #1438, Moved some Pgp.internal methods to other classes
